+++ src/UIRenderer.lua
@@ -53,6 +53,9 @@ UIRenderer.Render = function(renderer, node)
 	if node.data.eventManager then
 		node.data.eventManager:Resume()
 	end
+	if node.data.viewportSizeEventManager then
+		node.data.viewportSizeEventManager:Resume()
+	end
 end
 
 UIRenderer.Update = function(renderer, node, newElement)
@@ -61,6 +64,9 @@ UIRenderer.Update = function(renderer, node, newElement)
 
 	if node.data.eventManager then
 		node.data.eventManager:Suspend()
+	end
+	if node.data.viewportSizeEventManager then
+		node.data.viewportSizeEventManager:Suspend()
 	end
 
 	local name = node.key
@@ -91,6 +97,9 @@ UIRenderer.Update = function(renderer, node, newElement)
 
 	if node.data.eventManager then
 		node.data.eventManager:Resume()
+	end
+	if node.data.viewportSizeEventManager then
+		node.data.viewportSizeEventManager:Resume()
 	end
 
 	return node